✨ refactor and move transform, and utils to separate files
