+++ packages/eds-tokens-build/bin/eds-tokens-build.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import { run } from '../dist/index.js'
+import { run } from '../dist/eds-tokens-build.js'
 
 await run()
