+++ packages/eds-tokens-build/src/utils/index.ts
@@ -1,0 +1,43 @@
+export function toFixedWithoutTrailingZeroes(
+  value: number,
+  fractionDigits = 3,
+) {
+  const valueWithTwoDigitsAfterDecimal = value.toFixed(fractionDigits)
+  const valueWithoutTrailingZeroes = parseFloat(
+    `${valueWithTwoDigitsAfterDecimal}`,
+  )
+  return valueWithoutTrailingZeroes
+}
+
+export function transformNumberToRem(value: number): string {
+  return transformNumberToUnit(value, 16, 'rem')
+}
+
+function transformNumberToUnit(
+  value: number,
+  rootFontSize: number,
+  unit: string,
+): string {
+  if (value === 0) return `0${unit}`
+
+  const valueWithTwoDigitsAfterDecimal = Number(value.toFixed(3))
+  const valueInUnit = valueWithTwoDigitsAfterDecimal / rootFontSize
+  const valueWithoutTrailingZeroes = toFixedWithoutTrailingZeroes(valueInUnit)
+  const valueWithSuffix = `${valueWithoutTrailingZeroes}${unit}`
+
+  return valueWithSuffix
+}
+
+// manually convert reference into custom property
+export const resolveReference = (value: string, prefix: string): string => {
+  if (!value) return ''
+
+  const valueFormatted = value
+    .toLowerCase()
+    .replace('{', '')
+    .replace('}', '')
+    .replaceAll(' ', '-')
+    .replaceAll('.', '-')
+
+  return `var(--${prefix}-${valueFormatted})`
+}
